[変更]DL関係のクラスを細分化
